backend/scripts/requirements.txt
requests>=2.31.0
python-nmap>=0.7.1
beautifulsoup4>=4.12.0
urllib3>=2.0.0
dnspython>=2.4.0
cryptography>=41.0.0
lxml>=4.9.0
